Multiple hidden layers... glitchy
Multiple hidden layers actually seem to be working, but they're not right. 2 hidden layers is less correct than a single hidden layer using the example currently in the main.cpp file. They do work though, so this is close to the solution

I also changed the example in main.cpp to invert the input, instead of copy the input, to be a more advanced example.